+++ src/compas_timber/fabrication/btlx.py
@@ -225,7 +225,7 @@ class BTLxWriter(object):
         # Add part references if any beams are assigned to this stock
         if stock.element_data:
             for beam_guid, data in stock.element_data.items():
-                position_frame = data.get("frame", None)
+                position_frame = data.frame
                 # Apply scale factor to the frame
                 if scale_factor != 1.0:
                     position_frame = position_frame.scaled(scale_factor)